+++ pixell/lensing.py
@@ -76,22 +76,6 @@ def lens_map_flat(cmb_map, phi_map):
 
 ######## Curved sky lensing ########
 
-def almxfl(alm,lfunc,ainfo=None):
-	"""Filter alms isotropically by a function.
-	Returns alm * lfunc(ell)
-
-	Args:
-	    alm: (...,N) ndarray of spherical harmonic alms
-	    lfunc: a function mapping multipole ell to the filtering expression
-	    ainfo: 	If ainfo is provided, it is an alm_info describing the layout 
-	of the input alm. Otherwise it will be inferred from the alm itself.
-
-	Returns:
-	    falm: The filtered alms alm * lfunc(ell)
-	"""
-	ainfo = sharp.alm_info(nalm=alm.shape[-1]) if ainfo is None else ainfo
-	l = np.arange(ainfo.lmax+1.0)
-	return ainfo.lmul(alm, lfunc(l))
 
 def phi_to_kappa(phi_alm,phi_ainfo=None):
 	"""Convert lensing potential alms phi_alm to
@@ -106,7 +90,8 @@ def phi_to_kappa(phi_alm,phi_ainfo=None):
 	Returns:
 	    kappa_alm: The filtered alms phi_alm * l * (l+1) / 2
 	"""
-	return almxfl(alm=phi_alm,lfunc=lambda x: x*(x+1)/2,ainfo=phi_ainfo)
+	from . import curvedsky
+	return curvedsky.almxfl(alm=phi_alm,lfunc=lambda x: x*(x+1)/2,ainfo=phi_ainfo)
 
 def lens_map_curved(shape, wcs, phi_alm, cmb_alm, phi_ainfo=None, maplmax=None, dtype=np.float64, oversample=2.0, spin=[0,2], output="l", geodesic=True, verbose=False, delta_theta=None):
 	from . import curvedsky, sharp
